Found bugs, fixed rock to ground
When spamming NA the numbers don't reset correctly, need to change how combo's work.
Rocks sometimes glitch when they can not spawn due to altitude.
Rocks now stick to the ground when needed and otherwise can float in the air. Altered the animator location so that root motion can be used for some attacks later on (was quite a chore so hope it pays off)