material
set custom lit on clothes for doublesided and smoothness features.
Something went wrong with the anim controller for Shizen (Model.controller)... It had resetted to something it was in another branch for unknown reason. I managed to grab a good one from an older version, but I should save it elsewhere as well just in case